+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/header';
+import Header from '@/components/layout/header';
 import Hero from '@/components/section/hero';
 import About from '@/components/section/about';
 import Skills from '@/components/section/skills';
